+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 ## --- [4.11.0] - 2026/TBD
 ### New features
-TBD
+- Add OpenGraph social media previews/embeds link customization ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/1029))
+- Custom Branding and "Customize Crafty" page redesign ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/1037))
+- Resolve glob patterns in logfile paths for Hytale logs ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/1039))
 ### Bug fixes
 TBD
 ### Tweaks